requirements.txt
django
djangorestframework
pygments
ipython==8.7.0
